+++ deco/src/components/Common/AnswerEditor/AnswerEditor.module.css
@@ -31,48 +31,3 @@
   font-size: var(--text-sm);
   font-weight: 400;
 }
-
-/* displayState = false */
-.buttonWrapper {
-  display: none;
-  position: absolute;
-  top: 66px;
-  right: 27px;
-  background-color: var(--white);
-  border: 1px solid var(--gray-400);
-  border-radius: 8px;
-  padding: 17px;
-  font-size: var(--text-sm);
-  font-weight: 700;
-}
-
-/* displayState = true */
-.buttonWrapperDisplay {
-  position: absolute;
-  top: 66px;
-  right: 27px;
-  background-color: var(--white);
-  border: 1px solid var(--gray-400);
-  border-radius: 8px;
-  padding: 17px;
-  font-size: var(--text-sm);
-  font-weight: 700;
-}
-
-.buttonWrapper button {
-  display: flex;
-  border: none;
-  font-size: var(--text-sm);
-  font-weight: 700;
-  background-color: var(--white);
-}
-
-.buttonWrapper button:first-child {
-  margin-bottom: 17px;
-}
-
-.dotThree {
-  position: absolute;
-  top: 35px;
-  right: 30px;
-}
